Finish moving network code to networkClient
